Move EC2 imds helper to their own file
